Le choix des films est plus cohérent + Prise en charge des cas où TMDB n'est pas complet

+++ backend/src/main/java/cinequiz/backend/api_questions/tmdb_objects/MovieInfos.java
@@ -6,7 +6,7 @@ public class MovieInfos {
     public boolean adult;
     public String backdrop_path;
     public BelongsToCollection belongs_to_collection;
-    public int budget;
+    public long budget;
     public ArrayList<Genre> genres;
     public String homepage;
     public int id;
@@ -19,7 +19,7 @@ public class MovieInfos {
     public ArrayList<ProductionCompany> production_companies;
     public ArrayList<ProductionCountry> production_countries;
     public String release_date;
-    public int revenue;
+    public long revenue;
     public int runtime;
     public ArrayList<SpokenLanguage> spoken_languages;
     public String status;
